Add validation of zarr chunks for sharded stores
